Attempt to resolve the remaining broken links
An attempt to solve the remaining broken links on the compiled list.
4,14,15,17,18

I added a :ref: link as requested, but I am no sure if it will work properly because some :ref: links don't work properly locally.
Additionally, I wasn't sure exactly what to do for 17 and 18, but I put the text in double back ticks as requested .